refactor: move url method into urlFor function for easier testing and typing
