+++ src/pages/Product/Gallery/Gallery.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { urlFor } from '../../../client';
+import { urlFor } from '../../../lib/client';
 import './Gallery.scss';
 
 type Props = {
@@ -35,7 +35,7 @@ function Gallery({ images, thumbnails }: Props) {
                 className='thumbnails__list'
                 onMouseEnter={() => setImageIndex(idx)}
               >
-                <img src={urlFor(thumbnail).url()} alt='product thumbnail' />
+                <img src={urlFor(thumbnail)} alt='product thumbnail' />
               </li>
             );
           })}
@@ -50,7 +50,7 @@ function Gallery({ images, thumbnails }: Props) {
                 <li key={idx} className={setImageClassName(idx)}>
                   <img
                     className='product__image'
-                    src={urlFor(image).url()}
+                    src={urlFor(image)}
                     alt='product'
                     onMouseEnter={(e) => {
                       // update image size and turn-on magnifier
@@ -94,7 +94,7 @@ function Gallery({ images, thumbnails }: Props) {
           opacity: '1',
           border: '1px solid lightgray',
           backgroundColor: 'white',
-          backgroundImage: `url('${urlFor(images[imageIndex]).url()}')`,
+          backgroundImage: `url('${urlFor(images[imageIndex])}')`,
           backgroundRepeat: 'no-repeat',
 
           // calculate zoomed image size
